+++ src/app/features/android/store/android.effects.spec.ts
@@ -1,4 +1,4 @@
-import { buildTaskTitle, readableUrl } from './android.effects';
+import { buildTaskTitle } from './android.effects';
 
 describe('android share helpers', () => {
   describe('buildTaskTitle', () => {
@@ -66,22 +66,6 @@ describe('android share helpers', () => {
       });
       expect(result.length).toBe(150);
       expect(result.endsWith('...')).toBeTrue();
-    });
-  });
-
-  describe('readableUrl', () => {
-    it('returns host and decoded path', () => {
-      expect(readableUrl('https://www.example.com/foo/bar-baz')).toBe(
-        'example.com: foo bar baz',
-      );
-    });
-
-    it('returns just the host when there is no path', () => {
-      expect(readableUrl('https://example.com/')).toBe('example.com');
-    });
-
-    it('returns the original string for invalid URLs', () => {
-      expect(readableUrl('not a url')).toBe('not a url');
     });
   });
 });
