+++ src/app/features/android/store/android.effects.ts
@@ -8,6 +8,7 @@ import { androidInterface, AndroidShareData } from '../android-interface';
 import { TaskService } from '../../tasks/task.service';
 import { TaskAttachmentService } from '../../tasks/task-attachment/task-attachment.service';
 import { T } from '../../../t.const';
+import { readableUrl } from '../../../util/readable-url';
 
 // TODO send message to electron when current task changes here
 
@@ -204,25 +205,4 @@ export const buildTaskTitle = (shareData: Partial<AndroidShareData>): string => 
   }
 
   return taskTitle.length > 150 ? taskTitle.substring(0, 147) + '...' : taskTitle;
-};
-
-/**
- * Turn a URL into a human-readable "host: path" string for use as a task title.
- */
-export const readableUrl = (url: string): string => {
-  try {
-    const parsed = new URL(url);
-    const host = parsed.hostname.replace(/^www\./, '');
-    const pathPart = parsed.pathname.replace(/\/$/, '');
-    if (pathPart && pathPart !== '/') {
-      const decoded = decodeURIComponent(pathPart)
-        .replace(/[/_-]/g, ' ')
-        .replace(/\s+/g, ' ')
-        .trim();
-      return decoded ? `${host}: ${decoded}` : host;
-    }
-    return host;
-  } catch {
-    return url;
-  }
 };
